Adding check for node processes to avoid bash error

+++ stop_gzweb.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
 
-killall node
+if [ "$(pidof node)" ] 
+then
+  killall node
+fi
